backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "clerk-backend-api>=2.2.0",
    "fastapi>=0.115.12",
    "openai>=1.79.0",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.41",
    "svix>=1.65.0",
    "uvicorn>=0.34.2",
]
